Реализовано ограничение доступа к картам других пользователей. Отредактирован PackRestController.

+++ src/main/java/com/example/projectlaminam/rest/CardRestController.java
@@ -5,6 +5,8 @@ import com.example.projectlaminam.domain.Pack;
 import com.example.projectlaminam.repositories.CardRepository;
 import com.example.projectlaminam.repositories.PackRepository;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 import java.net.URISyntaxException;
@@ -26,23 +28,41 @@ public class CardRestController {
     }
 
     @GetMapping
-    public List<Card> getCards(@PathVariable("packId") Long packId) throws RuntimeException {
+    public ResponseEntity<List<Card>> getCards(@PathVariable("packId") Long packId,
+                               @AuthenticationPrincipal UserDetails user) throws RuntimeException {
 
         Pack pack = packRepository.findById(packId).orElseThrow(RuntimeException::new);
-        return cardRepository.findAllByPack(pack);
+        if(pack.getUsers().stream().noneMatch(u -> u.getUsername().equals(user.getUsername()))) {
+            return ResponseEntity.status(403).build();
+        }
+        return ResponseEntity.ok(cardRepository.findAllByPack(pack));
     }
 
     @PostMapping
-    public ResponseEntity<Card> createCard(@RequestBody Card card, @PathVariable("packId") Long packId) throws URISyntaxException, RuntimeException {
+    public ResponseEntity<Card> createCard(@RequestBody Card card,
+                                           @PathVariable("packId") Long packId,
+                                           @AuthenticationPrincipal UserDetails user) throws URISyntaxException, RuntimeException {
 
         Pack currentPack = packRepository.findById(packId).orElseThrow(RuntimeException::new);
+        if(currentPack.getUsers().stream().noneMatch(u -> u.getUsername().equals(user.getUsername()))) {
+            return ResponseEntity.status(403).build();
+        }
+
         card.setPack(currentPack);
         Card savedCard = cardRepository.save(card);
         return ResponseEntity.ok(savedCard);
     }
 
     @DeleteMapping("/{cardId}")
-    public ResponseEntity<Card> deleteCard(@PathVariable("cardId") Long cardId) {
+    public ResponseEntity<Card> deleteCard(@PathVariable("packId") Long packId,
+                                           @PathVariable("cardId") Long cardId,
+                                           @AuthenticationPrincipal UserDetails user) {
+
+        Pack currentPack = packRepository.findById(packId).orElseThrow(RuntimeException::new);
+        if(currentPack.getUsers().stream().noneMatch(u -> u.getUsername().equals(user.getUsername()))) {
+            return ResponseEntity.status(403).build();
+        }
+
         cardRepository.deleteById(cardId);
         return ok().build();
     }
@@ -50,16 +70,23 @@ public class CardRestController {
     @PutMapping("/{cardId}")
     public ResponseEntity<Card> updateCard(@RequestBody Card card,
                                            @PathVariable("cardId") Long cardId,
-                                           @PathVariable("packId") Long packId) {
+                                           @PathVariable("packId") Long packId,
+                                           @AuthenticationPrincipal UserDetails user) {
 
-        if (card != null && Objects.equals(card.getId(), cardId)) {
-            Card currentCard = cardRepository.findById(card.getId()).orElseThrow(RuntimeException::new);
-            currentCard.setLabel(card.getLabel());
-            currentCard.setFrontSide(card.getFrontSide());
-            currentCard.setBackSide(card.getBackSide());
-            currentCard = cardRepository.save(currentCard);
-            return ResponseEntity.ok(currentCard);
+        Pack currentPack = packRepository.findById(packId).orElseThrow(RuntimeException::new);
+        if(currentPack.getUsers().stream().noneMatch(u -> u.getUsername().equals(user.getUsername()))) {
+            return ResponseEntity.status(403).build();
+        }else if (card == null || !Objects.equals(card.getId(), cardId)) {
+            return ResponseEntity.notFound().build();
         }
-        return ResponseEntity.notFound().build();
+
+        Card currentCard = cardRepository.findById(card.getId()).orElseThrow(RuntimeException::new);
+        currentCard.setLabel(card.getLabel());
+        currentCard.setFrontSide(card.getFrontSide());
+        currentCard.setBackSide(card.getBackSide());
+        currentCard = cardRepository.save(currentCard);
+
+        return ResponseEntity.ok(currentCard);
+
     }
 }
